Code refactoring, now flags can be of arbitrary types.
The macro `@symbol_flagset` corresponds to the old `@flagset` macro.
Now `@flagset` evaluates the provided flags and a different syntax is needed
for custom bit values.
The old syntax "flag = bit" still works if `FlagType` is `Symbol` of if
`::BaseType` is used.
If some `UndefVarError` is thrown, an useful error message is printed,
notifying the user of the incompatible syntax.